# argos-service-api/api.yml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Argos Service
  license:
    name: Apache License 2.0
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /supplychain/{supplyChainId}/layout:
    get:
      summary: get layout
      operationId: getLayout
      tags:
        - layout
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutMetaBlock"
        '500':
          $ref: '#/components/responses/Error'
        '404':
          description: no layout found
    post:
      summary: create or update a layout
      operationId: createOrUpdateLayout
      tags:
        - layout
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayoutMetaBlock"
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutMetaBlock"
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created layout metablock can be found
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
  /supplychain/{supplyChainId}/layout/validate:
    post:
      summary: validates a layout on data input and model consistency
      operationId: validateLayout
      tags:
        - layout
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Layout"
      responses:
        '204':
          description: layout is correct
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
  /supplychain/{supplyChainId}/layout/releaseconfig:
    get:
      summary: returns release configurations for a layout
      operationId: getReleaseConfiguration
      tags:
        - layout
      security:
        - bearerAuth: []
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: release configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseConfiguration"
        '500':
          $ref: '#/components/responses/Error'
    post:
      summary: create release configurations for a layout
      operationId: createReleaseConfiguration
      tags:
        - layout
      security:
        - bearerAuth: []
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseConfiguration"
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseConfiguration"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
  /supplychain/{supplyChainId}/layout/approvalconfig:
    get:
      summary: returns a list of approval configurations for a layout
      operationId: getApprovalConfigurations
      tags:
        - layout
      security:
        - bearerAuth: []
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: array of approval configurations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalConfiguration"
        '500':
          $ref: '#/components/responses/Error'
    post:
      summary: create approval configurations for a layout
      operationId: createApprovalConfigurations
      tags:
        - layout
      security:
        - bearerAuth: []
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 20
              items:
                $ref: "#/components/schemas/ApprovalConfiguration"
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalConfiguration"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
  /supplychain/{supplyChainId}/layout/approvalconfig/me:
    get:
      summary: returns a list of approvals for the login user
      operationId: getApprovalsForAccount
      tags:
        - layout
      security:
        - bearerAuth: []
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: array of approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalConfiguration"
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    ValidationError:
      description: validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Error:
      description: server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    HexString:
      type: string
      maxLength: 1024
      pattern: '^[0-9a-f]*$'
    SHA256Hash:
      type: string
      minLength: 64
      maxLength: 64
      pattern: '^[0-9a-f]*$'
    UUID:
      type: string
      description: an Argos Service identifier
      minLength: 36
      maxLength: 36
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}?$'
    UnixPath:
      type: string
      maxLength: 1024
      pattern: '^(?!.*\\).*$'
    NameString:
      type: string
      minLength: 1
      maxLength: 63
      pattern: '^([a-z]|[a-z][a-z0-9-]*[a-z0-9])?$'
    ByteString:
      type: string
      maxLength: 4096
      format: byte
...
    Signature:
      type: object
      required:
        - keyId
        - signature
        - hashAlgorithm
        - keyAlgorithm
      properties:
        keyId:
          $ref: '#/components/schemas/KeyId'
        signature:
          $ref: "#/components/schemas/HexString"
        hashAlgorithm:
          $ref: '#/components/schemas/HashAlgorithm'
        keyAlgorithm:
          $ref: '#/components/schemas/KeyAlgorithm'
...
    PublicKey:
      type: object
      required:
        - keyId
        - publicKey
      properties:
        keyId:
          $ref: '#/components/schemas/KeyId'
        publicKey:
          $ref: '#/components/schemas/ByteString'
...
    LayoutMetaBlock:
      type: object
      required:
        - signatures
        - layout
      properties:
        signatures:
          minItems: 1
          maxItems: 20
          type: array
          items:
            $ref: "#/components/schemas/Signature"
        layout:
          $ref: "#/components/schemas/Layout"
    Layout:
      type: object
      required:
        - keys
        - authorizedKeyIds
        - expectedEndProducts
        - layoutSegments
      properties:
        keys:
          type: array
          minItems: 1
          maxItems: 256
          items:
            $ref: '#/components/schemas/PublicKey'
        authorizedKeyIds:
          type: array
          minItems: 1
          maxItems: 256
          items:
            $ref: "#/components/schemas/SHA256Hash"
        expectedEndProducts:
          type: array
          minItems: 1
          maxItems: 4096
          items:
            $ref: '#/components/schemas/MatchRule'
        layoutSegments:
          type: array
          minItems: 1
          maxItems: 256
          items:
            $ref: "#/components/schemas/LayoutSegment"
    LayoutSegment:
      type: object
      required:
        - name
        - steps
      properties:
        name:
          $ref: "#/components/schemas/NameString"
        steps:
          type: array
          minItems: 1
          maxItems: 256
          items:
            $ref: "#/components/schemas/Step"
    MatchRule:
      type: object
      required:
        - pattern
        - destinationType
        - destinationSegmentName
        - destinationStepName
      properties:
        pattern:
          type: string
          maxLength: 1024
        sourcePathPrefix:
          $ref: "#/components/schemas/UnixPath"
        destinationType:
          enum:
            - PRODUCTS
            - MATERIALS
        destinationPathPrefix:
          $ref: "#/components/schemas/UnixPath"
        destinationSegmentName:
          $ref: "#/components/schemas/NameString"
        destinationStepName:
          $ref: "#/components/schemas/NameString"
    Step:
      type: object
      required:
        - name
        - requiredNumberOfLinks
        - authorizedKeyIds
        - expectedMaterials
        - expectedProducts
      properties:
        name:
          $ref: "#/components/schemas/NameString"
        authorizedKeyIds:
          type: array
          minItems: 1
          maxItems: 256
          items:
            $ref: '#/components/schemas/KeyId'
        requiredNumberOfLinks:
          type: integer
          format: int32
          minimum: 1
          maximum: 20
        expectedMaterials:
          type: array
          maxItems: 4096
          items:
            $ref: "#/components/schemas/Rule"
        expectedProducts:
          type: array
          maxItems: 4096
          items:
            $ref: "#/components/schemas/Rule"
    KeyId:
      $ref: "#/components/schemas/SHA256Hash"
    KeyAlgorithm:
      enum:
        - EC
    HashAlgorithm:
      enum:
        - SHA256
        - SHA384
        - SHA512
    Rule:
      type: object
      required:
        - ruleType
        - pattern
      properties:
        ruleType:
          enum:
            - ALLOW
            - CREATE
            - CREATE_OR_MODIFY
            - DELETE
            - DISALLOW
            - MATCH
            - MODIFY
            - REQUIRE
        pattern:
          type: string
          maxLength: 1024
        sourcePathPrefix:
          $ref: "#/components/schemas/UnixPath"
        destinationPathPrefix:
          $ref: "#/components/schemas/UnixPath"
        destinationSegmentName:
          $ref: "#/components/schemas/NameString"
        destinationType:
          enum:
            - PRODUCTS
            - MATERIALS
        destinationStepName:
          $ref: "#/components/schemas/NameString"
...
    ReleaseConfiguration:
      type: object
      required:
        - artifactCollectorSpecifications
      properties:
        artifactCollectorSpecifications:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ArtifactCollectorSpecification"
    ApprovalConfiguration:
      type: object
      required:
        - stepName
        - segmentName
        - artifactCollectorSpecifications
      properties:
        stepName:
          $ref: "#/components/schemas/NameString"
        segmentName:
          $ref: "#/components/schemas/NameString"
        artifactCollectorSpecifications:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ArtifactCollectorSpecification"
          minItems: 1
    ArtifactCollectorSpecification:
      type: object
      required:
        - name
        - type
        - uri
        - context
      properties:
        name:
          $ref: "#/components/schemas/NameString"
        type:
          enum:
            - XLDEPLOY
            - GIT
        uri:
          description: base url of the artifact collector server
          type: string
          format: uri
        context:
          description: additional key value pair attributes that can very per collector type for XLDEPLOY the key is 'applicationName'
          type: object
          additionalProperties:
            type: string
            maxLength: 1024
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
    ValidationMessage:
      type: object
      required:
        - message
        - type
      properties:
        field:
          type: string
        type:
          enum:
            - DATA_INPUT
            - MODEL_CONSISTENCY
        message:
          type: string
...